+++ lib/Formula/Formula.cpp
@@ -1,6 +1,6 @@
 #include "Formula/Formula.h"
 
-Formula::Formula( int numberOfDice, int maxValueOfDice, int levelMultiplier, int constant ) 
+Formula::Formula( int numberOfDice, int maxValueOfDice, float levelMultiplier, int constant ) 
 		: dice( numberOfDice, maxValueOfDice ), levelMultiplier(levelMultiplier), constant(constant) 
 {
